PI-3622 refactor GetRegion.kt & tidy
