+++ projects/accredited-programmes-and-delius/src/dev/kotlin/uk/gov/justice/digital/hmpps/data/DataLoader.kt
@@ -120,7 +120,6 @@ class DataLoader(
     private val officeLocationRepository: OfficeLocationRepository,
     private val providerRepository: ProviderRepository,
     private val enforcementActionRepository: EnforcementActionRepository,
-    private val pduRepository: PduRepository,
 ) : ApplicationListener<ApplicationReadyEvent> {
     @PostConstruct
     fun saveAuditUser() {
@@ -200,7 +199,5 @@ class DataLoader(
         contactRepository.save(TWO_THIRDS_CONTACT)
         contactRepository.save(OTHER_CONTACT)
         contactRepository.saveAll(APPOINTMENTS)
-        pduRepository.save(PDU)
-        pduRepository.save(PDU_2)
     }
 }
